Add Automation for AWS and SSH keys
This isnt really idempotent but its a start... the way the logic works
is that a groovy script is dropped that is run on every jenkins boot. If details
change it isnt smart enough right now to update or delete missing keys. Will
try to mix that logic in later. Or maybe we should be running dynamically
against the /cli tool...... I dunno, this is good enough for now :+1: