Be more aggressive with updating cloud config
I was a little conservative earlier with not trying to clobber existing
cloud configuration. This was a mistake since certain parameters of cloud
config have been changing rapidly during testing. This change makes it so
if there is a cloud config with the same name as another one, it will always
clobber on jenkins reboot.