Add 'google-compute-engine' configuration support
Added configuration templating for setup of the 'google-compute-engine'.
This includes adding OAuth json keys in the credentials store and
setting up the builder options on the Google "cloud" section.

There are two main issues that still need to be addressed:

* This commit adds a potential issue with jep200 which I had to add a
java cli work-around for. I'm a groovy n00b so I'm sure its something
easy. Punting for now though.
* I didn't feel like exposing the gpu configuration options via ansible.
I'm sure its easy to do but I didn't need it so :shrug: