GCE Cloud configuration fixes
Turns out the previous config I laid out was ... seemingly getting the
Google Cloud configured but it left a lot of broken gaps. There has to
be a much more dynamic way to do this.. but for now .. I'm hardcoding
most of the API URLs in.

One problem here is I removed the configurability of the network
settings. That will need to get fixed later.