Ensure that existing creds get updated on reboot
This updates the previous groovy credential injection logic to actually
update existing entries .. so that for instance passwords and
descriptions can be modified in-place. Note that if you update the id
name of a credential entry, the logic wont be able to update an existing
entry to match that and instead a new entry is created. So if you do
that, all job references to that key also have to be updated.

Inspired from
https://github.com/carlosfrias/configure-jenkins-workstation/blob/081b8c9fc7bf8713eee651cc8b83fc3b9dde184d/jenkins-docker-config/jenkins-init/credentials.groovy

Source code at:
https://github.com/jenkinsci/credentials-plugin/blob/master/src/main/java/com/cloudbees/plugins/credentials/SystemCredentialsProvider.java#L593